[Exporter] Refactor code, added more tests for uncovered resources
Split code into more files to make work with `importables.go` a bit easier.

NO_CHANGELOG=true
