+++ exporter/impl_uc.go
@@ -10,6 +10,7 @@ import (
 	"github.com/databricks/databricks-sdk-go/service/catalog"
 	"github.com/databricks/databricks-sdk-go/service/dataquality"
 	"github.com/databricks/databricks-sdk-go/service/tags"
+	"github.com/databricks/databricks-sdk-go/service/vectorsearch"
 	tf_uc "github.com/databricks/terraform-provider-databricks/catalog"
 	"github.com/databricks/terraform-provider-databricks/common"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -719,6 +720,94 @@ func listDataQualityMonitors(ic *importContext) error {
 			Resource: "databricks_data_quality_monitor",
 			ID:       fmt.Sprintf("%s,%s", monitor.ObjectType, monitor.ObjectId),
 		})
+	}
+	return nil
+}
+
+func listVectorSearchEndpoints(ic *importContext) error {
+	endpoints, err := ic.workspaceClient.VectorSearchEndpoints.ListEndpointsAll(ic.Context, vectorsearch.ListEndpointsRequest{})
+	if err != nil {
+		log.Printf("[ERROR] listing vector search endpoints: %s", err.Error())
+		return err
+	}
+	for _, ep := range endpoints {
+		ic.EmitIfUpdatedAfterMillisAndNameMatches(&resource{
+			Resource: "databricks_vector_search_endpoint",
+			ID:       ep.Name,
+		}, ep.Name, ep.LastUpdatedTimestamp, fmt.Sprintf("vector search endpoint '%s'", ep.Name))
+	}
+	return nil
+}
+
+func importVectorSearchEndpoint(ic *importContext, r *resource) error {
+	indexes, err := ic.workspaceClient.VectorSearchIndexes.ListIndexesAll(ic.Context, vectorsearch.ListIndexesRequest{
+		EndpointName: r.ID,
+	})
+	if err != nil {
+		log.Printf("[ERROR] listing vector search indexes for endpoint %s: %s", r.ID, err.Error())
+		return err
+	}
+	for _, idx := range indexes {
+		ic.Emit(&resource{
+			Resource: "databricks_vector_search_index",
+			ID:       idx.Name,
+		})
+	}
+	return nil
+}
+
+func importVectorSearchIndex(ic *importContext, r *resource) error {
+	ic.emitUCGrantsWithOwner("table/"+r.ID, r)
+	s := ic.Resources["databricks_vector_search_index"].Schema
+	var vsi vectorsearch.VectorIndex
+	common.DataToStructPointer(r.Data, s, &vsi)
+	if vsi.EndpointName != "" {
+		ic.Emit(&resource{
+			Resource: "databricks_vector_search_endpoint",
+			ID:       vsi.EndpointName,
+		})
+	}
+	if vsi.DeltaSyncIndexSpec != nil {
+		ic.Emit(&resource{
+			Resource: "databricks_sql_table",
+			ID:       vsi.DeltaSyncIndexSpec.SourceTable,
+		})
+		if vsi.DeltaSyncIndexSpec.EmbeddingWritebackTable != "" {
+			ic.Emit(&resource{
+				Resource: "databricks_sql_table",
+				ID:       vsi.DeltaSyncIndexSpec.EmbeddingWritebackTable,
+			})
+		}
+		for _, col := range vsi.DeltaSyncIndexSpec.EmbeddingSourceColumns {
+			if col.EmbeddingModelEndpointName != "" {
+				ic.Emit(&resource{
+					Resource: "databricks_model_serving",
+					ID:       col.EmbeddingModelEndpointName,
+				})
+			}
+			if col.ModelEndpointNameForQuery != "" {
+				ic.Emit(&resource{
+					Resource: "databricks_model_serving",
+					ID:       col.ModelEndpointNameForQuery,
+				})
+			}
+		}
+	}
+	if vsi.DirectAccessIndexSpec != nil {
+		for _, col := range vsi.DirectAccessIndexSpec.EmbeddingSourceColumns {
+			if col.EmbeddingModelEndpointName != "" {
+				ic.Emit(&resource{
+					Resource: "databricks_model_serving",
+					ID:       col.EmbeddingModelEndpointName,
+				})
+			}
+			if col.ModelEndpointNameForQuery != "" {
+				ic.Emit(&resource{
+					Resource: "databricks_model_serving",
+					ID:       col.ModelEndpointNameForQuery,
+				})
+			}
+		}
 	}
 	return nil
 }
