+++ exporter/impl_workspace.go
@@ -3,6 +3,7 @@ package exporter
 import (
 	"encoding/base64"
 	"fmt"
+	"log"
 	"strconv"
 	"strings"
 
@@ -84,5 +85,98 @@ func analyzeNotebook(ic *importContext, content string) {
 			}
 		}
 	}
+}
 
+func searchRepoByPath(ic *importContext, r *resource) error {
+	repoDir, err := ic.workspaceClient.Workspace.GetStatusByPath(ic.Context, r.Value)
+	if err != nil {
+		return err
+	}
+	if repoDir.ObjectType != sdk_workspace.ObjectTypeRepo {
+		return fmt.Errorf("object %s is not a repo", r.Value)
+	}
+	if repoDir.ResourceId != "" {
+		r.ID = repoDir.ResourceId
+	} else {
+		r.ID = strconv.FormatInt(repoDir.ObjectId, 10)
+	}
+	return nil
+}
+
+func listRepos(ic *importContext) error {
+	it := ic.workspaceClient.Repos.List(ic.Context, sdk_workspace.ListReposRequest{PathPrefix: "/Workspace"})
+	i := 1
+	for it.HasNext(ic.Context) {
+		repo, err := it.Next(ic.Context)
+		if err != nil {
+			return err
+		}
+		if !ic.MatchesName(repo.Path) {
+			log.Printf("[INFO] Repo %s doesn't match %s filter", repo.Path, ic.match)
+			continue
+		}
+		if repo.Url != "" {
+			ic.Emit(&resource{
+				Resource: "databricks_repo",
+				ID:       strconv.FormatInt(repo.Id, 10),
+			})
+		} else {
+			log.Printf("[WARN] ignoring databricks_repo without Git provider. Path: %s", repo.Path)
+			ic.addIgnoredResource(fmt.Sprintf("databricks_repo. path=%s", repo.Path))
+		}
+		if i%50 == 0 {
+			log.Printf("[INFO] Scanned %d repos", i)
+		}
+		i++
+	}
+	return nil
+}
+
+func importRepo(ic *importContext, r *resource) error {
+	path := maybeStripWorkspacePrefix(r.Data.Get("path").(string))
+	if strings.HasPrefix(path, "/Repos") {
+		ic.emitUserOrServicePrincipalForPath(path, "/Repos")
+	} else if strings.HasPrefix(path, "/Users") {
+		ic.emitUserOrServicePrincipalForPath(path, "/Users")
+	}
+	ic.emitPermissionsIfNotIgnored(r, fmt.Sprintf("/repos/%s", r.ID),
+		"repo_"+ic.Importables["databricks_repo"].Name(ic, r.Data))
+	return nil
+}
+
+func importWorkspaceFile(ic *importContext, r *resource) error {
+	ic.emitUserOrServicePrincipalForPath(r.ID, "/Users")
+	resp, err := ic.workspaceClient.Workspace.Export(ic.Context, sdk_workspace.ExportRequest{
+		Path:   r.ID,
+		Format: sdk_workspace.ExportFormatAuto,
+	})
+	if err != nil {
+		if apierr.IsMissing(err) {
+			ic.addIgnoredResource(fmt.Sprintf("databricks_workspace_file. path=%s", r.ID))
+		}
+		return err
+	}
+	objectId := r.Data.Get("object_id").(int)
+	parts := strings.Split(r.ID, "/")
+	plen := len(parts)
+	if idx := strings.Index(parts[plen-1], "."); idx != -1 {
+		parts[plen-1] = parts[plen-1][:idx] + "_" + strconv.Itoa(objectId) + parts[plen-1][idx:]
+	} else {
+		parts[plen-1] = parts[plen-1] + "_" + strconv.Itoa(objectId)
+	}
+	name := fileNameNormalizationRegex.ReplaceAllString(strings.Join(parts, "/")[1:], "_")
+	content, _ := base64.StdEncoding.DecodeString(resp.Content)
+	fileName, err := ic.saveFileIn("workspace_files", name, []byte(content))
+	if err != nil {
+		return err
+	}
+
+	ic.emitPermissionsIfNotIgnored(r, fmt.Sprintf("/files/%d", objectId),
+		"ws_file_"+ic.Importables["databricks_workspace_file"].Name(ic, r.Data))
+
+	// TODO: it's not completely correct condition - we need to make emit smarter -
+	// emit only if permissions are different from their parent's permission.
+	ic.emitWorkspaceObjectParentDirectory(r)
+	log.Printf("[TRACE] Creating %s for %s", fileName, r)
+	return r.Data.Set("source", fileName)
 }
