+++ exporter/importables_test.go
@@ -2,7 +2,6 @@ package exporter
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"log"
 	"os"
@@ -11,35 +10,20 @@ import (
 	"testing"
 
 	"github.com/databricks/databricks-sdk-go/apierr"
-	sdk_uc "github.com/databricks/databricks-sdk-go/service/catalog"
+	"github.com/databricks/databricks-sdk-go/listing"
 	"github.com/databricks/databricks-sdk-go/service/compute"
 	"github.com/databricks/databricks-sdk-go/service/iam"
-	sdk_jobs "github.com/databricks/databricks-sdk-go/service/jobs"
-	sdk_pipelines "github.com/databricks/databricks-sdk-go/service/pipelines"
-	sdk_sharing "github.com/databricks/databricks-sdk-go/service/sharing"
-	sdk_vs "github.com/databricks/databricks-sdk-go/service/vectorsearch"
 	sdk_workspace "github.com/databricks/databricks-sdk-go/service/workspace"
-	tf_uc "github.com/databricks/terraform-provider-databricks/catalog"
-	"github.com/databricks/terraform-provider-databricks/clusters"
 	"github.com/databricks/terraform-provider-databricks/commands"
 	"github.com/databricks/terraform-provider-databricks/common"
 	"github.com/databricks/terraform-provider-databricks/internal/providers/sdkv2"
-	tf_jobs "github.com/databricks/terraform-provider-databricks/jobs"
 	"github.com/databricks/terraform-provider-databricks/permissions"
 	"github.com/databricks/terraform-provider-databricks/permissions/entity"
-	tf_dlt "github.com/databricks/terraform-provider-databricks/pipelines"
-	"github.com/databricks/terraform-provider-databricks/policies"
-	"github.com/databricks/terraform-provider-databricks/pools"
 	"github.com/databricks/terraform-provider-databricks/qa"
-	"github.com/databricks/terraform-provider-databricks/repos"
 	"github.com/databricks/terraform-provider-databricks/scim"
 	"github.com/databricks/terraform-provider-databricks/secrets"
-	tf_settings "github.com/databricks/terraform-provider-databricks/settings"
-	tf_sharing "github.com/databricks/terraform-provider-databricks/sharing"
 	"github.com/databricks/terraform-provider-databricks/storage"
-	tf_vs "github.com/databricks/terraform-provider-databricks/vectorsearch"
 	"github.com/databricks/terraform-provider-databricks/workspace"
-	"github.com/hashicorp/hcl/v2/hclwrite"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/exp/maps"
@@ -93,132 +77,30 @@ func importContextForTestWithClient(ctx context.Context, client *common.Databric
 	return ic
 }
 
-func TestInstancePool(t *testing.T) {
-	d := pools.ResourceInstancePool().ToResource().TestResourceData()
-	d.Set("instance_pool_name", "blah-bah")
-	ic := importContextForTest()
-	ic.enableServices("access,pools")
-
-	name := resourcesMap["databricks_instance_pool"].Name(ic, d)
-	assert.Equal(t, "blah-bah", name)
-
-	d = pools.ResourceInstancePool().ToResource().TestResourceData()
-	d.SetId("abc-bcd-def")
-	name = resourcesMap["databricks_instance_pool"].Name(ic, d)
-	assert.Equal(t, "def", name)
-
-	ic.meAdmin = true
-	r := &resource{
-		ID:   "abc",
-		Data: d,
-	}
-	err := resourcesMap["databricks_instance_pool"].Import(ic, r)
-	assert.NoError(t, err)
-	assert.True(t, ic.testEmits["databricks_permissions[inst_pool_def] (id: /instance-pools/abc)"])
+// importContextForAccountTestWithClient creates an import context configured for account-level testing
+func importContextForAccountTestWithClient(ctx context.Context, client *common.DatabricksClient, services string) *importContext {
+	client.Config.AccountID = testAccountID
+	client.Config.WithTesting()
+	ic := importContextForTestWithClient(ctx, client)
+	ic.enableServices(services)
+	return ic
 }
 
-func TestClusterPolicy(t *testing.T) {
-	d := policies.ResourceClusterPolicy().ToResource().TestResourceData()
-	d.Set("name", "bcd")
-	definition := map[string]map[string]string{
-		"aws_attributes.instance_profile_arn": {
-			"value": "def",
+// Helper function to create an iterator from a slice
+func createIteratorFromSlice[T any](items []T) listing.Iterator[T] {
+	request := struct{}{}
+	return listing.NewIterator(
+		&request,
+		func(ctx context.Context, req struct{}) ([]T, error) {
+			return items, nil
 		},
-		"instance_pool_id": {
-			"defaultValue": "efg",
+		func(resp []T) []T {
+			return resp
 		},
-		"init_scripts.0.dbfs.destination": {
-			"type":  "fixed",
-			"value": "dbfs:/FileStore/init-script.sh",
+		func(resp []T) *struct{} {
+			return nil // No pagination
 		},
-	}
-	policy, _ := json.Marshal(definition)
-	d.Set("definition", string(policy))
-	ic := importContextForTest()
-	ic.enableServices("storage,pools,policies,access")
-	ic.meAdmin = true
-	err := ic.Importables["databricks_cluster_policy"].Import(ic, &resource{
-		ID:   "abc",
-		Data: d,
-	})
-	assert.NoError(t, err)
-	assert.Len(t, ic.testEmits, 4)
-	assert.True(t, ic.testEmits["databricks_permissions[cluster_policy_bcd] (id: /cluster-policies/abc)"])
-	assert.True(t, ic.testEmits["databricks_instance_pool[<unknown>] (id: efg)"])
-	assert.True(t, ic.testEmits["databricks_instance_profile[<unknown>] (id: def)"])
-	assert.True(t, ic.testEmits["databricks_dbfs_file[<unknown>] (id: dbfs:/FileStore/init-script.sh)"])
-}
-
-func TestPredefinedClusterPolicy(t *testing.T) {
-	d := policies.ResourceClusterPolicy().ToResource().TestResourceData()
-	d.Set("policy_family_id", "job-cluster")
-	d.Set("name", "Job Compute")
-	policy, _ := json.Marshal(map[string]map[string]string{})
-	d.Set("definition", string(policy))
-	ic := importContextForTest()
-	ic.builtInPolicies = map[string]compute.PolicyFamily{
-		"job-cluster": {Name: "Job Compute"},
-	}
-	r := resource{ID: "abc", Data: d}
-	err := ic.Importables["databricks_cluster_policy"].Import(ic, &r)
-	assert.NoError(t, err)
-	assert.Equal(t, "data", r.Mode)
-	assert.Equal(t, "", r.Data.Get("definition").(string))
-}
-
-func TestGroup(t *testing.T) {
-	ic := importContextForTest()
-	ic.enableServices("groups,access")
-	ic.allGroups = []scim.Group{
-		{
-			DisplayName: "foo",
-			ID:          "123",
-			Roles: []scim.ComplexValue{
-				{
-					Value: "abc",
-					Type:  "direct",
-				},
-			},
-			Members: []scim.ComplexValue{
-				// this is just for log line printing
-				{Value: "a001"},
-				{Value: "a002"},
-				{Value: "a003"},
-				{Value: "a004"},
-				{Value: "a005"},
-				{Value: "a006"},
-				{Value: "a007"},
-				{Value: "a008"},
-				{Value: "a009"},
-				{Value: "a010"},
-				{Value: "a011"},
-			},
-			Groups: []scim.ComplexValue{
-				{
-					Value: "parent-group",
-					Type:  "direct",
-				},
-			},
-		},
-	}
-	d := scim.ResourceGroup().ToResource().TestResourceData()
-	d.Set("display_name", "foo")
-	r := &resource{
-		Value:     "foo",
-		Attribute: "display_name",
-		Data:      d,
-	}
-	err := ic.Importables["databricks_group"].Search(ic, r)
-	assert.NoError(t, err)
-	assert.Equal(t, "123", r.ID)
-
-	err = ic.Importables["databricks_group"].Import(ic, r)
-	assert.NoError(t, err)
-	assert.Len(t, ic.testEmits, 4)
-	assert.True(t, ic.testEmits["databricks_group_role[<unknown>] (id: 123|abc)"])
-	assert.True(t, ic.testEmits["databricks_instance_profile[<unknown>] (id: abc)"])
-	assert.True(t, ic.testEmits["databricks_group[<unknown>] (id: parent-group)"])
-	assert.True(t, ic.testEmits["databricks_group_member[_parent-group_foo] (id: parent-group|123)"])
+	)
 }
 
 func TestPermissions(t *testing.T) {
@@ -264,536 +146,6 @@ func TestSecretScope(t *testing.T) {
 	assert.Equal(t, "abc_a9993e3647", name)
 }
 
-func TestInstancePoolNameFromID(t *testing.T) {
-	ic := importContextForTest()
-	d := pools.ResourceInstancePool().ToResource().TestResourceData()
-	d.SetId("a-b-c")
-	d.Set("instance_pool_name", "")
-	assert.Equal(t, "c", resourcesMap["databricks_instance_pool"].Name(ic, d))
-}
-
-func TestClusterNameFromID(t *testing.T) {
-	ic := importContextForTest()
-	d := clusters.ResourceCluster().ToResource().TestResourceData()
-	d.SetId("a-b-c")
-	assert.Equal(t, "c", resourcesMap["databricks_cluster"].Name(ic, d))
-}
-
-func TestRepoName(t *testing.T) {
-	ic := importContextForTest()
-	d := repos.ResourceRepo().ToResource().TestResourceData()
-	d.SetId("12345")
-	// Repo without path
-	assert.Equal(t, "repo_12345", resourcesMap["databricks_repo"].Name(ic, d))
-	// Repo with path
-	d.Set("path", "/Repos/user/test")
-	assert.Equal(t, "user_test_12345", resourcesMap["databricks_repo"].Name(ic, d))
-}
-
-func TestRepoIgnore(t *testing.T) {
-	ic := importContextForTest()
-	d := repos.ResourceRepo().ToResource().TestResourceData()
-	d.SetId("12345")
-	d.Set("path", "/Repos/user/test")
-	r := &resource{ID: "12345", Data: d}
-	// Repo without URL
-	assert.True(t, resourcesMap["databricks_repo"].Ignore(ic, r))
-	assert.Equal(t, 1, len(ic.ignoredResources))
-	// Repo with URL
-	d.Set("url", "https://github.com/abc/abc.git")
-	assert.False(t, resourcesMap["databricks_repo"].Ignore(ic, r))
-}
-
-func TestDLTIgnore(t *testing.T) {
-	ic := importContextForTest()
-	d := tf_dlt.ResourcePipeline().ToResource().TestResourceData()
-	scm := tf_dlt.ResourcePipeline().Schema
-
-	d.SetId("12345")
-	r := &resource{ID: "12345", Data: d}
-	// job without libraries
-	assert.True(t, resourcesMap["databricks_pipeline"].Ignore(ic, r))
-	assert.Equal(t, 1, len(ic.ignoredResources))
-
-	// job deployed by DABs
-	d.MarkNewResource()
-	pipeline := tf_dlt.Pipeline{
-		PipelineSpec: sdk_pipelines.PipelineSpec{
-			Deployment: &sdk_pipelines.PipelineDeployment{
-				Kind: "BUNDLE",
-			},
-		},
-	}
-	err := common.StructToData(pipeline, scm, d)
-	require.NoError(t, err)
-
-	r = &resource{ID: "12345", Data: d}
-	for k := range ic.ignoredResources {
-		delete(ic.ignoredResources, k)
-	}
-	assert.True(t, resourcesMap["databricks_pipeline"].Ignore(ic, r))
-	assert.Equal(t, 1, len(ic.ignoredResources))
-}
-
-func TestJobsIgnore(t *testing.T) {
-	ic := importContextForTest()
-	d := tf_jobs.ResourceJob().ToResource().TestResourceData()
-	d.SetId("12345")
-	r := &resource{ID: "12345", Data: d}
-	// job without tasks
-	assert.True(t, resourcesMap["databricks_job"].Ignore(ic, r))
-	assert.Equal(t, 1, len(ic.ignoredResources))
-}
-
-func TestJobName(t *testing.T) {
-	ic := importContextForTest()
-	d := tf_jobs.ResourceJob().ToResource().TestResourceData()
-	d.SetId("12345")
-	// job without name
-	assert.Equal(t, "job_12345", resourcesMap["databricks_job"].Name(ic, d))
-	// job with name
-	d.Set("name", "test@1pm")
-	assert.Equal(t, "test_1pm_12345", resourcesMap["databricks_job"].Name(ic, d))
-}
-
-func TestImportClusterLibraries(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Status:       200,
-			Resource:     "/api/2.0/libraries/cluster-status?cluster_id=abc",
-			Response: compute.ClusterLibraryStatuses{
-				LibraryStatuses: []compute.LibraryFullStatus{
-					{
-						Library: &compute.Library{
-							Whl: "foo.whl",
-						},
-						Status: "INSTALLED",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		d := clusters.ResourceCluster().ToResource().TestResourceData()
-		d.SetId("abc")
-		err := resourcesMap["databricks_cluster"].Import(ic, &resource{
-			ID:   "abc",
-			Data: d,
-		})
-		assert.NoError(t, err)
-	})
-}
-
-func TestImportClusterLibrariesFails(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Status:       404,
-			Resource:     "/api/2.0/libraries/cluster-status?cluster_id=abc",
-			Response: &apierr.APIError{
-				ErrorCode:  "NOT_FOUND",
-				StatusCode: 404,
-				Message:    "nope",
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		d := clusters.ResourceCluster().ToResource().TestResourceData()
-		d.SetId("abc")
-		err := resourcesMap["databricks_cluster"].Import(ic, &resource{
-			ID:   "abc",
-			Data: d,
-		})
-		assert.EqualError(t, err, "nope")
-	})
-}
-
-func TestClusterListFails(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.1/clusters/list?filter_by.cluster_sources=UI&filter_by.cluster_sources=API&page_size=100",
-			Status:   404,
-			Response: &apierr.APIError{
-				ErrorCode:  "NOT_FOUND",
-				StatusCode: 404,
-				Message:    "nope",
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		err := resourcesMap["databricks_cluster"].List(ic)
-		assert.EqualError(t, err, "nope")
-	})
-}
-
-func TestClusterList_NoNameMatch(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.1/clusters/list?filter_by.cluster_sources=UI&filter_by.cluster_sources=API&page_size=100",
-			Response: clusters.ClusterList{
-				Clusters: []clusters.ClusterInfo{
-					{
-						ClusterName: "abc",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.match = "bcd"
-		err := resourcesMap["databricks_cluster"].List(ic)
-		assert.NoError(t, err)
-		assert.Equal(t, 0, len(ic.testEmits))
-	})
-}
-
-func TestInstnacePoolsListWithMatch(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/instance-pools/list",
-			Response: compute.ListInstancePools{
-				InstancePools: []compute.InstancePoolAndStats{
-					{
-						InstancePoolName: "test",
-						InstancePoolId:   "test",
-					},
-					{
-						InstancePoolName: "bcd",
-						InstancePoolId:   "bcd",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("pools")
-		ic.match = "bcd"
-		err := resourcesMap["databricks_instance_pool"].List(ic)
-		assert.NoError(t, err)
-		assert.Equal(t, 1, len(ic.testEmits))
-	})
-}
-
-func TestJobListNoNameMatchOrFromBundles(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.2/jobs/list?limit=100",
-			Response: sdk_jobs.ListJobsResponse{
-				Jobs: []sdk_jobs.BaseJob{
-					{
-						Settings: &sdk_jobs.JobSettings{
-							Name: "abc",
-						},
-					},
-					{
-						Settings: &sdk_jobs.JobSettings{
-							Name:     "bcd",
-							EditMode: "UI_LOCKED",
-							Deployment: &sdk_jobs.JobDeployment{
-								Kind: "BUNDLE",
-							},
-						},
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("jobs")
-		ic.match = "bcd"
-		err := resourcesMap["databricks_job"].List(ic)
-		assert.NoError(t, err)
-		assert.Equal(t, 0, len(ic.testEmits))
-	})
-}
-
-func TestClusterPolicyWrongDef(t *testing.T) {
-	d := policies.ResourceClusterPolicy().ToResource().TestResourceData()
-	d.Set("name", "abc")
-	d.Set("definition", "..")
-	ic := importContextForTest()
-	err := resourcesMap["databricks_cluster_policy"].Import(ic, &resource{
-		ID:   "x",
-		Data: d,
-	})
-	assert.EqualError(t, err, "invalid character '.' looking for beginning of value")
-}
-
-func TestClusterPolicyNoValues(t *testing.T) {
-	d := policies.ResourceClusterPolicy().ToResource().TestResourceData()
-	d.Set("name", "abc")
-	d.Set("definition", `{"foo": {}}`)
-	ic := importContextForTest()
-	ic.enableServices("access")
-	ic.meAdmin = true
-	err := resourcesMap["databricks_cluster_policy"].Import(ic, &resource{
-		ID:   "x",
-		Data: d,
-	})
-	assert.NoError(t, err)
-	assert.Equal(t, 1, len(ic.testEmits))
-}
-
-func TestGroupCacheAndSearchError(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Resource:     "/api/2.0/preview/scim/v2/Groups?attributes=id&count=10000&startIndex=1",
-			Status:       404,
-			Response: &apierr.APIError{
-				ErrorCode:  "NOT_FOUND",
-				StatusCode: 404,
-				Message:    "nope",
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("groups")
-		err := resourcesMap["databricks_group"].List(ic)
-		assert.EqualError(t, err, "nope")
-
-		err = resourcesMap["databricks_group"].Search(ic, &resource{
-			Attribute: "display_name",
-		})
-		assert.EqualError(t, err, "nope")
-
-		err = resourcesMap["databricks_group"].Search(ic, &resource{
-			Attribute: "nonsense",
-		})
-		assert.EqualError(t, err, "wrong search attribute 'nonsense' for databricks_group")
-		d := scim.ResourceGroup().ToResource().TestResourceData()
-		d.Set("display_name", "nonsense")
-		err = resourcesMap["databricks_group"].Import(ic, &resource{
-			ID:   "nonsense",
-			Data: d,
-		})
-		assert.EqualError(t, err, "nope")
-	})
-}
-
-func TestGroupListNoNameMatch(t *testing.T) {
-	ic := importContextForTest()
-	ic.match = "bcd"
-	ic.allGroups = []scim.Group{
-		{
-			DisplayName: "abc",
-		},
-	}
-	err := resourcesMap["databricks_group"].List(ic)
-	assert.NoError(t, err)
-	assert.Equal(t, 0, len(ic.testEmits))
-}
-
-func TestGroupSearchNoMatch(t *testing.T) {
-	ic := importContextForTest()
-	ic.allGroups = []scim.Group{
-		{
-			DisplayName: "abc",
-		},
-	}
-	r := &resource{
-		Attribute: "display_name",
-		Value:     "dbc",
-	}
-	err := resourcesMap["databricks_group"].Search(ic, r)
-	assert.NoError(t, err)
-	assert.Equal(t, "", r.ID)
-}
-
-func TestUserSearchFails(t *testing.T) {
-	userFixture := qa.ListUsersFixtures([]iam.User{})
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		userFixture[0],
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		d := scim.ResourceUser().ToResource().TestResourceData()
-		d.Set("user_name", "dbc")
-		r := &resource{
-			Attribute: "display_name",
-			Value:     "dbc",
-			Data:      d,
-		}
-		err := resourcesMap["databricks_user"].Search(ic, r)
-		assert.EqualError(t, err, "there is no user 'dbc'")
-
-		err = resourcesMap["databricks_user"].Import(ic, r)
-		assert.EqualError(t, err, "user dbc is not found")
-	})
-}
-
-func TestSpnSearchFails(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		qa.ListServicePrincipalsFixtures([]iam.ServicePrincipal{})[0],
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		d := scim.ResourceServicePrincipal().ToResource().TestResourceData()
-		d.Set("application_id", "dbc")
-		r := &resource{
-			Attribute: "application_id",
-			Value:     "dbc",
-			Data:      d,
-		}
-		err := resourcesMap["databricks_service_principal"].Search(ic, r)
-		assert.EqualError(t, err, "there is no service principal 'dbc'")
-
-		err = resourcesMap["databricks_service_principal"].Import(ic, r)
-		assert.EqualError(t, err, "service principal dbc is not found")
-	})
-}
-
-func TestSpnSearchSuccess(t *testing.T) {
-	spFixture := qa.ListServicePrincipalsFixtures([]iam.ServicePrincipal{
-		{
-			Id: "321", DisplayName: "spn", ApplicationId: "dbc",
-		},
-	})
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		spFixture[0],
-		spFixture[1],
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Resource:     "/api/2.0/preview/scim/v2/ServicePrincipals/321?attributes=userName,displayName,active,externalId,entitlements,groups,roles",
-			Response:     scim.User{ID: "321", DisplayName: "spn", ApplicationID: "dbc"},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		d := scim.ResourceServicePrincipal().ToResource().TestResourceData()
-		d.Set("application_id", "dbc")
-		d.Set("display_name", "dbc")
-		r := &resource{
-			Attribute: "application_id",
-			Value:     "dbc",
-			Data:      d,
-		}
-		err := resourcesMap["databricks_service_principal"].Search(ic, r)
-		assert.NoError(t, err)
-
-		err = resourcesMap["databricks_service_principal"].Import(ic, r)
-		assert.NoError(t, err)
-
-		assert.True(t, resourcesMap["databricks_service_principal"].ShouldOmitField(ic, "application_id",
-			scim.ResourceServicePrincipal().Schema["application_id"], d, r))
-		ic.Client.Config.Host = "https://abc.azuredatabricks.net"
-		// We shouldn't omit display_name for Databricks-managed SPs
-		assert.False(t, resourcesMap["databricks_service_principal"].ShouldOmitField(ic, "display_name",
-			scim.ResourceServicePrincipal().Schema["display_name"], d, r))
-		assert.True(t, resourcesMap["databricks_service_principal"].ShouldOmitField(ic, "application_id",
-			scim.ResourceServicePrincipal().Schema["application_id"], d, r))
-		// We shouldn't omit application_id for Azure-managed SPs, but omit display_name
-		d.Set("external_id", "60622399-fd3f-4faf-8810-bf08b225cf3b")
-		assert.False(t, resourcesMap["databricks_service_principal"].ShouldOmitField(ic, "application_id",
-			scim.ResourceServicePrincipal().Schema["application_id"], d, r))
-		assert.True(t, resourcesMap["databricks_service_principal"].ShouldOmitField(ic, "display_name",
-			scim.ResourceServicePrincipal().Schema["display_name"], d, r))
-
-		// test for different branches in Name function
-		// test for different branches in Name function
-		d2 := scim.ResourceServicePrincipal().ToResource().TestResourceData()
-		d2.SetId("123")
-		d2.Set("application_id", "dbc")
-		assert.Equal(t, "dbc_123", resourcesMap["databricks_service_principal"].Name(ic, d2))
-		d2.Set("application_id", "60622399-fd3f-4faf-8810-bf08b225cf3b")
-		assert.Equal(t, "60622399_123", resourcesMap["databricks_service_principal"].Name(ic, d2))
-
-		d2.Set("display_name", "abc")
-		assert.Equal(t, "abc_123", resourcesMap["databricks_service_principal"].Name(ic, d2))
-	})
-}
-
-func TestShouldOmitForUsers(t *testing.T) {
-	d := scim.ResourceUser().ToResource().TestResourceData()
-	d.SetId("user1")
-	d.Set("user_name", "user@domain.com")
-	d.Set("display_name", "")
-	r := &resource{
-		Attribute: "databricks_user",
-		Value:     "user@domain.com",
-		Data:      d,
-	}
-	assert.True(t, resourcesMap["databricks_user"].ShouldOmitField(nil, "display_name",
-		scim.ResourceUser().Schema["display_name"], d, r))
-	d.Set("display_name", "user@domain.com")
-	assert.True(t, resourcesMap["databricks_user"].ShouldOmitField(nil, "display_name",
-		scim.ResourceUser().Schema["display_name"], d, r))
-	d.Set("display_name", "Some user")
-	assert.False(t, resourcesMap["databricks_user"].ShouldOmitField(nil, "display_name",
-		scim.ResourceUser().Schema["display_name"], d, r))
-}
-
-func TestShouldOmitFoRepos(t *testing.T) {
-	d := repos.ResourceRepo().ToResource().TestResourceData()
-	d.SetId("1234")
-	d.Set("path", "/Repos/Test/repo")
-	r := &resource{
-		Attribute: "databricks_repo",
-		Value:     "repo",
-		Data:      d,
-	}
-	assert.False(t, resourcesMap["databricks_repo"].ShouldOmitField(nil, "path",
-		repos.ResourceRepo().Schema["path"], d, r))
-	assert.True(t, resourcesMap["databricks_repo"].ShouldOmitField(nil, "branch",
-		repos.ResourceRepo().Schema["branch"], d, r))
-	assert.True(t, resourcesMap["databricks_repo"].ShouldOmitField(nil, "tag",
-		repos.ResourceRepo().Schema["tag"], d, r))
-	d.Set("branch", "test")
-	assert.False(t, resourcesMap["databricks_repo"].ShouldOmitField(nil, "branch",
-		repos.ResourceRepo().Schema["branch"], d, r))
-	d.Set("tag", "v123")
-	assert.False(t, resourcesMap["databricks_repo"].ShouldOmitField(nil, "tag",
-		repos.ResourceRepo().Schema["tag"], d, r))
-}
-
-func TestUserImportSkipNonDirectGroups(t *testing.T) {
-	userFixture := qa.ListUsersFixtures([]iam.User{
-		{
-			UserName: "dbc",
-			Id:       "321",
-		},
-	})
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		userFixture[0],
-		userFixture[1],
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Resource:     "/api/2.0/preview/scim/v2/Users/321?attributes=id,userName,displayName,active,externalId,entitlements,groups,roles",
-			Response: scim.UserList{
-				Resources: []scim.User{
-					{
-						Groups: []scim.ComplexValue{
-							{
-								Display: "x",
-								Value:   "y",
-							},
-						},
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		d := scim.ResourceUser().ToResource().TestResourceData()
-		d.Set("user_name", "dbc")
-		r := &resource{
-			Attribute: "display_name",
-			Value:     "dbc",
-			Data:      d,
-		}
-		err := resourcesMap["databricks_user"].Import(ic, r)
-		assert.NoError(t, err)
-		assert.Equal(t, 0, len(ic.testEmits))
-	})
-}
-
 func TestSecretScopeListNoNameMatch(t *testing.T) {
 	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
 		{
@@ -815,179 +167,6 @@ func TestSecretScopeListNoNameMatch(t *testing.T) {
 		assert.NoError(t, err)
 		assert.Equal(t, 0, len(ic.testEmits))
 	})
-}
-
-func TestPoliciesListing(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/policies/clusters/list?",
-			Response: compute.ListPoliciesResponse{
-				Policies: []compute.Policy{
-					{
-						Name:           "Personal Compute",
-						PolicyFamilyId: "personal-vm",
-						PolicyId:       "123",
-					},
-					{
-						Name:     "abcd",
-						PolicyId: "456",
-					},
-				},
-			},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/policy-families?",
-			Response: compute.ListPolicyFamiliesResponse{
-				PolicyFamilies: []compute.PolicyFamily{
-					{
-						Name:           "Personal Compute",
-						PolicyFamilyId: "personal-vm",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("policies")
-		err := resourcesMap["databricks_cluster_policy"].List(ic)
-		assert.NoError(t, err)
-		assert.Equal(t, 1, len(ic.testEmits))
-	})
-}
-
-func TestPoliciesListNoNameMatch(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/policies/clusters/list?",
-			Response: compute.ListPoliciesResponse{
-				Policies: []compute.Policy{
-					{
-						Name: "Personal Compute",
-					},
-					{
-						Name: "abcd",
-					},
-				},
-			},
-		},
-		emptyPolicyFamilies,
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("policies")
-		ic.match = "bcd"
-		err := resourcesMap["databricks_cluster_policy"].List(ic)
-		assert.NoError(t, err)
-		assert.Equal(t, 0, len(ic.testEmits))
-	})
-}
-
-func TestAwsS3MountProfile(t *testing.T) {
-	ic := importContextForTest()
-	ic.mounts = true
-	ic.match = "abc"
-	ic.enableServices("mounts,access")
-	ic.mountMap = map[string]mount{}
-	ic.mountMap["/mnt/abc"] = mount{
-		URL:             "s3a://abc",
-		InstanceProfile: "bcd",
-	}
-	ic.mountMap["/mnt/def"] = mount{
-		URL:             "s3a://def",
-		InstanceProfile: "bcd",
-	}
-	err := resourcesMap["databricks_mount"].List(ic)
-	assert.NoError(t, err)
-	assert.Len(t, ic.testEmits, 2)
-	assert.True(t, ic.testEmits["databricks_instance_profile[<unknown>] (id: bcd)"])
-	assert.True(t, ic.testEmits["databricks_mount[<unknown>] (id: /mnt/abc)"])
-}
-
-func TestMountsBodyGeneration(t *testing.T) {
-	ic := importContextForTest()
-	ic.mounts = true
-	ic.match = "abc"
-	ic.mountMap = map[string]mount{}
-	ic.variables = map[string]string{}
-	ic.mountMap["/mnt/abc"] = mount{
-		URL:             "s3a://abc",
-		InstanceProfile: "bcd",
-	}
-	ic.mountMap["/mnt/def"] = mount{
-		URL:       "s3a://def",
-		ClusterID: "bcd",
-	}
-	ic.mountMap["/mnt/gcs"] = mount{
-		URL:       "gs://gcs/dir",
-		ClusterID: "bcd",
-	}
-	ic.mountMap["/mnt/abfss"] = mount{
-		URL: "abfss://test@test.dfs.core.windows.net/directory",
-	}
-	ic.mountMap["/mnt/wasbs"] = mount{
-		URL: "wasbs://test@test.blob.core.windows.net/directory",
-	}
-	ic.mountMap["/mnt/adls"] = mount{
-		URL: "adls://test.dfs.core.windows.net/directory",
-	}
-	ic.mountMap["/mnt/dbfs"] = mount{
-		URL: "dbfs:/directory",
-	}
-
-	//
-	f := hclwrite.NewEmptyFile()
-	body := f.Body()
-
-	err := generateMountBody(ic, body, &resource{
-		ID:       "/mnt/abc",
-		Name:     "abc",
-		Resource: "databricks_mount",
-	})
-	assert.NoError(t, err)
-
-	err = generateMountBody(ic, body, &resource{
-		ID:       "/mnt/def",
-		Name:     "def",
-		Resource: "databricks_mount",
-	})
-	assert.NoError(t, err)
-
-	err = generateMountBody(ic, body, &resource{
-		ID:       "/mnt/abfss",
-		Name:     "abfss",
-		Resource: "databricks_mount",
-	})
-	assert.NoError(t, err)
-
-	err = generateMountBody(ic, body, &resource{
-		ID:       "/mnt/gcs",
-		Name:     "gcs",
-		Resource: "databricks_mount",
-	})
-	assert.NoError(t, err)
-
-	err = generateMountBody(ic, body, &resource{
-		ID:       "/mnt/adls",
-		Name:     "adls",
-		Resource: "databricks_mount",
-	})
-	assert.NoError(t, err)
-
-	err = generateMountBody(ic, body, &resource{
-		ID:       "/mnt/wasbs",
-		Name:     "wasbs",
-		Resource: "databricks_mount",
-	})
-	assert.NoError(t, err)
-
-	err = generateMountBody(ic, body, &resource{
-		ID:       "/mnt/dbfs",
-		Name:     "dbfs",
-		Resource: "databricks_mount",
-	})
-	assert.EqualError(t, err, "no matching handler for: dbfs:/directory")
 }
 
 func TestGlobalInitScriptNameFromId(t *testing.T) {
@@ -1053,53 +232,6 @@ func TestGlobalInitScriptsBodyErrors(t *testing.T) {
 	})
 }
 
-func TestRepoListFails(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			MatchAny:     true,
-			Status:       404,
-			Response: &apierr.APIError{
-				ErrorCode:  "NOT_FOUND",
-				StatusCode: 404,
-				Message:    "nope",
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		err := resourcesMap["databricks_repo"].List(ic)
-		assert.EqualError(t, err, "nope")
-	})
-}
-
-func TestNotebookWorkspaceFileImportNotFound(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			MatchAny:     true,
-			Status:       404,
-			Response: &apierr.APIError{
-				ErrorCode:  "NOT_FOUND",
-				StatusCode: 404,
-				Message:    "nope",
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		err := resourcesMap["databricks_notebook"].Import(ic, &resource{
-			ID: "/abc",
-		})
-		assert.EqualError(t, err, "nope")
-		assert.Contains(t, ic.ignoredResources, "databricks_notebook. path=/abc")
-
-		err = resourcesMap["databricks_workspace_file"].Import(ic, &resource{
-			ID: "/def",
-		})
-		assert.EqualError(t, err, "nope")
-		assert.Contains(t, ic.ignoredResources, "databricks_workspace_file. path=/def")
-	})
-}
-
 func testGenerate(t *testing.T, fixtures []qa.HTTPFixture, services string, asAdmin bool, cb func(*importContext)) {
 	qa.HTTPFixturesApply(t, fixtures, func(ctx context.Context, client *common.DatabricksClient) {
 		ic := importContextForTestWithClient(ctx, client)
@@ -1124,200 +256,6 @@ func getGeneratedFile(ic *importContext, service string) string {
 		return ""
 	}
 	return string(content)
-}
-
-func TestNotebookGeneration(t *testing.T) {
-	testGenerate(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/workspace/list?path=%2F",
-			Response: workspace.ObjectList{
-				Objects: []workspace.ObjectStatus{
-					{
-						Path:       "/Repos/Foo/Bar",
-						ObjectType: "NOTEBOOK",
-					},
-					{
-						Path:       "/First/Second",
-						ObjectType: "NOTEBOOK",
-						ObjectID:   123,
-						Language:   "PYTHON",
-					},
-				},
-			},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/workspace/export?format=SOURCE&path=%2FFirst%2FSecond",
-			Response: workspace.ExportPath{
-				Content: "YWJj",
-			},
-			ReuseRequest: true,
-		},
-	}, "notebooks", false, func(ic *importContext) {
-		ic.notebooksFormat = "SOURCE"
-		ic.enableListing("notebooks")
-		err := listWorkspaceObjects(ic)
-		assert.NoError(t, err)
-		ic.waitGroup.Wait()
-		ic.closeImportChannels()
-		ic.generateAndWriteResources(nil)
-		assert.Equal(t, commands.TrimLeadingWhitespace(`
-		resource "databricks_notebook" "first_second_123" {
-		  source = "${path.module}/notebooks/First/Second_123.py"
-		  path   = "/First/Second"
-		}`), getGeneratedFile(ic, "notebooks"))
-	})
-}
-
-func TestNotebookGenerationJupyter(t *testing.T) {
-	testGenerate(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/workspace/list?path=%2F",
-			Response: workspace.ObjectList{
-				Objects: []workspace.ObjectStatus{
-					{
-						Path:       "/Repos/Foo/Bar",
-						ObjectType: "NOTEBOOK",
-					},
-					{
-						ObjectID:   123,
-						ObjectType: "NOTEBOOK",
-						Path:       "/First/Second",
-						Language:   "PYTHON",
-					},
-				},
-			},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/workspace/export?format=JUPYTER&path=%2FFirst%2FSecond",
-			Response: workspace.ExportPath{
-				Content: "YWJj",
-			},
-			ReuseRequest: true,
-		},
-	}, "notebooks", false, func(ic *importContext) {
-		ic.notebooksFormat = "JUPYTER"
-		ic.enableListing("notebooks")
-		err := listWorkspaceObjects(ic)
-		assert.NoError(t, err)
-		ic.waitGroup.Wait()
-		ic.closeImportChannels()
-		ic.generateAndWriteResources(nil)
-		assert.Equal(t, commands.TrimLeadingWhitespace(`
-		resource "databricks_notebook" "first_second_123" {
-		  source   = "${path.module}/notebooks/First/Second_123.ipynb"
-		  path     = "/First/Second"
-		  language = "PYTHON"
-		  format   = "JUPYTER"
-		}`), getGeneratedFile(ic, "notebooks"))
-	})
-}
-
-func TestNotebookGenerationBadCharacters(t *testing.T) {
-	testGenerate(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/workspace/list?path=%2F",
-			Response: workspace.ObjectList{
-				Objects: []workspace.ObjectStatus{
-					{
-						Path:       "/Repos/Foo/Bar",
-						ObjectType: "NOTEBOOK",
-					},
-					{
-						ObjectID:   123,
-						ObjectType: "NOTEBOOK",
-						Path:       "/Fir\"st\\/Second",
-						Language:   "PYTHON",
-					},
-				},
-			},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/workspace/get-status?path=%2FFir%22st%5C",
-
-			Response: workspace.ObjectStatus{
-				ObjectID:   124,
-				ObjectType: "DIRECTORY",
-				Path:       "/Fir\"st\\",
-			},
-			ReuseRequest: true,
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/workspace/export?format=SOURCE&path=%2FFir%22st%5C%2FSecond",
-			Response: workspace.ExportPath{
-				Content: "YWJj",
-			},
-			ReuseRequest: true,
-		},
-	}, "notebooks,directories", true, func(ic *importContext) {
-		ic.notebooksFormat = "SOURCE"
-		ic.enableServices("notebooks")
-		ic.enableListing("notebooks")
-		err := listWorkspaceObjects(ic)
-		assert.NoError(t, err)
-		ic.waitGroup.Wait()
-		ic.closeImportChannels()
-		ic.generateAndWriteResources(nil)
-		assert.Equal(t, commands.TrimLeadingWhitespace(`
-		resource "databricks_notebook" "fir_st_second_123" {
-		  source = "${path.module}/notebooks/Fir_st_/Second_123.py"
-		  path   = "/Fir\"st\\/Second"
-		}`), getGeneratedFile(ic, "notebooks"))
-	})
-}
-
-func TestDirectoryGeneration(t *testing.T) {
-	testGenerate(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/workspace/list?path=%2F",
-			Response: workspace.ObjectList{
-				Objects: []workspace.ObjectStatus{
-					{
-						ObjectID:   1234,
-						Path:       "/first",
-						ObjectType: "DIRECTORY",
-					},
-				},
-			},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/workspace/list?path=%2Ffirst",
-			Response: workspace.ObjectList{
-				Objects: []workspace.ObjectStatus{
-					{},
-				},
-			},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/workspace/get-status?path=%2Ffirst",
-			Response: workspace.ObjectStatus{
-				ObjectID:   1234,
-				ObjectType: "DIRECTORY",
-				Path:       "/first",
-			},
-		},
-	}, "directories", false, func(ic *importContext) {
-		ic.enableListing("directories")
-		err := listWorkspaceObjects(ic)
-		assert.NoError(t, err)
-
-		ic.waitGroup.Wait()
-		ic.closeImportChannels()
-		ic.generateAndWriteResources(nil)
-		assert.Equal(t, commands.TrimLeadingWhitespace(`
-		resource "databricks_directory" "first_1234" {
-		  path = "/first"
-		}`), getGeneratedFile(ic, "directories"))
-	})
 }
 
 func TestGlobalInitScriptGeneration(t *testing.T) {
@@ -1381,40 +319,6 @@ func TestSecretGeneration(t *testing.T) {
 	})
 }
 
-func TestDbfsFileGeneration(t *testing.T) {
-	testGenerate(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/dbfs/get-status?path=a",
-			Response: storage.FileInfo{
-				Path: "a",
-			},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/dbfs/read?length=1000000&path=a",
-			Response: storage.ReadResponse{
-				Data:      "YWJj",
-				BytesRead: 3,
-			},
-		},
-	}, "storage", false, func(ic *importContext) {
-		ic.Emit(&resource{
-			Resource: "databricks_dbfs_file",
-			ID:       "a",
-		})
-
-		ic.waitGroup.Wait()
-		ic.closeImportChannels()
-		ic.generateAndWriteResources(nil)
-		assert.Equal(t, commands.TrimLeadingWhitespace(`
-		resource "databricks_dbfs_file" "_0cc175b9c0f1b6a831c399e269772661_a" {
-		  source = "${path.module}/dbfs_files/_0cc175b9c0f1b6a831c399e269772661_a"
-		  path   = "a"
-		}`), getGeneratedFile(ic, "storage"))
-	})
-}
-
 // TODO: remove it completely after we remove support for legacy dashboards
 func TestSqlListObjects(t *testing.T) {
 	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
@@ -1436,158 +340,6 @@ func TestSqlListObjects(t *testing.T) {
 		assert.NoError(t, err)
 		assert.Len(t, answer, 2)
 	})
-}
-
-func TestIncrementalListDLT(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/pipelines?max_results=100",
-			Response: sdk_pipelines.ListPipelinesResponse{
-				Statuses: []sdk_pipelines.PipelineStateInfo{
-					{
-						PipelineId: "abc",
-						Name:       "abc",
-					},
-					{
-						PipelineId: "def",
-						Name:       "def",
-					},
-				},
-			},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/pipelines/abc?",
-			Response: sdk_pipelines.GetPipelineResponse{
-				PipelineId:   "abc",
-				Name:         "abc",
-				LastModified: 1681466931226,
-			},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/pipelines/def?",
-			Response: sdk_pipelines.GetPipelineResponse{
-				PipelineId:   "def",
-				Name:         "def",
-				LastModified: 1690156900000,
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("dlt")
-		ic.incremental = true
-		ic.updatedSinceStr = "2023-07-24T00:00:00Z"
-		ic.updatedSinceMs = 1690156700000
-		err := resourcesMap["databricks_pipeline"].List(ic)
-		ic.waitGroup.Wait()
-		ic.closeImportChannels()
-		assert.NoError(t, err)
-		assert.Equal(t, 1, len(ic.testEmits))
-	})
-}
-
-func TestListSystemSchemasSuccess(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		currentMetastoreSuccess,
-		{
-			Method:   "GET",
-			Resource: fmt.Sprintf("/api/2.1/unity-catalog/metastores/%s/systemschemas?", currentMetastoreResponse.MetastoreId),
-			Response: sdk_uc.ListSystemSchemasResponse{
-				Schemas: []sdk_uc.SystemSchemaInfo{
-					{
-						Schema: "access",
-						State:  "ENABLE_COMPLETED",
-					},
-					{
-						Schema: "information_schema",
-						State:  "ENABLE_COMPLETED",
-					},
-					{
-						Schema: "marketplace",
-						State:  "available",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("uc-system-schemas")
-		ic.currentMetastore = currentMetastoreResponse
-		err := resourcesMap["databricks_system_schema"].List(ic)
-		assert.NoError(t, err)
-		assert.Equal(t, len(ic.testEmits), 1)
-	})
-}
-
-func TestListSystemSchemasErrorGetMetastore(t *testing.T) {
-	ic := importContextForTest()
-	err := resourcesMap["databricks_system_schema"].List(ic)
-	assert.EqualError(t, err, "there is no UC metastore information")
-}
-
-func TestListSystemSchemasErrorListing(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: fmt.Sprintf("/api/2.1/unity-catalog/metastores/%s/systemschemas?", currentMetastoreResponse.MetastoreId),
-			Status:   404,
-			Response: &apierr.APIError{
-				ErrorCode:  "NOT_FOUND",
-				StatusCode: 404,
-				Message:    "nope",
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.currentMetastore = currentMetastoreResponse
-		err := resourcesMap["databricks_system_schema"].List(ic)
-		assert.EqualError(t, err, "nope")
-	})
-}
-
-func TestListUcAllowListError(t *testing.T) {
-	ic := importContextForTest()
-	err := resourcesMap["databricks_artifact_allowlist"].List(ic)
-	assert.EqualError(t, err, "there is no UC metastore information")
-}
-
-func TestListUcAllowListSuccess(t *testing.T) {
-	ic := importContextForTest()
-	ic.enableServices("uc-artifact-allowlist")
-	ic.currentMetastore = currentMetastoreResponse
-	err := resourcesMap["databricks_artifact_allowlist"].List(ic)
-	assert.NoError(t, err)
-	assert.Equal(t, len(ic.testEmits), 3)
-	// Test ignore function
-	d := tf_uc.ResourceArtifactAllowlist().ToResource().TestResourceData()
-	d.MarkNewResource()
-	d.Set("id", "abc")
-	res := ic.Importables["databricks_artifact_allowlist"].Ignore(ic, &resource{
-		ID:   "abc",
-		Data: d,
-	})
-	assert.True(t, res)
-	assert.Contains(t, ic.ignoredResources, "databricks_artifact_allowlist. id=abc")
-	// Test ignore function, with blocks
-	err = common.StructToData(
-		tf_uc.ArtifactAllowlistInfo{
-			ArtifactType: "INIT_SCRIPT",
-			ArtifactMatchers: []sdk_uc.ArtifactMatcher{
-				{
-					Artifact:  "/Volumes/inits",
-					MatchType: "PREFIX_MATCH",
-				},
-			},
-		},
-		tf_uc.ResourceArtifactAllowlist().Schema, d)
-	assert.NoError(t, err)
-	res = ic.Importables["databricks_artifact_allowlist"].Ignore(ic, &resource{
-		ID:   "abc",
-		Data: d,
-	})
-	assert.False(t, res)
 }
 
 func TestEmitSqlParent(t *testing.T) {
@@ -1630,757 +382,6 @@ func TestEmitFilesFromSlice(t *testing.T) {
 		assert.Contains(t, ic.testEmits, "databricks_workspace_file[<unknown>] (id: /Shared/test.txt)")
 		assert.Contains(t, ic.testEmits, "databricks_repo[<unknown>] (id: 1234)")
 	})
-}
-
-func TestEmitFilesFromMap(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/workspace/get-status?path=%2FShared%2Ftest.txt&return_git_info=true",
-			Response: workspace.ObjectStatus{},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/workspace/get-status?path=%2FShared%2Fgit%2Ftest.txt&return_git_info=true",
-			Response: workspace.ObjectStatus{
-				GitInfo: &sdk_workspace.RepoInfo{
-					Id: 1234,
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("storage,notebooks,wsfiles,repos")
-		ic.emitFilesFromMap(map[string]string{
-			"k1": "dbfs:/FileStore/test.txt",
-			"k2": "/Workspace/Shared/test.txt",
-			"k3": "nothing",
-			"k4": "/Workspace/Shared/git/test.txt",
-		})
-		assert.Equal(t, 3, len(ic.testEmits))
-		assert.Contains(t, ic.testEmits, "databricks_dbfs_file[<unknown>] (id: dbfs:/FileStore/test.txt)")
-		assert.Contains(t, ic.testEmits, "databricks_workspace_file[<unknown>] (id: /Shared/test.txt)")
-		assert.Contains(t, ic.testEmits, "databricks_repo[<unknown>] (id: 1234)")
-	})
-}
-
-func TestStorageCredentialListFails(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.1/unity-catalog/storage-credentials?",
-			Status:   200,
-			Response: &sdk_uc.ListStorageCredentialsResponse{},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		err := resourcesMap["databricks_storage_credential"].List(ic)
-		assert.NoError(t, err)
-	})
-}
-
-func TestImportStorageCredentialGrants(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Status:       200,
-			Resource:     "/api/2.1/unity-catalog/permissions/storage_credential/abc",
-			Response: sdk_uc.GetPermissionsResponse{
-				PrivilegeAssignments: []sdk_uc.PrivilegeAssignment{
-					{
-						Principal:  "principal",
-						Privileges: []sdk_uc.Privilege{"CREATE EXTERNAL LOCATION"},
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		d := tf_uc.ResourceStorageCredential().ToResource().TestResourceData()
-		d.SetId("abc")
-		err := resourcesMap["databricks_storage_credential"].Import(ic, &resource{
-			ID:   "abc",
-			Data: d,
-		})
-		assert.NoError(t, err)
-	})
-}
-
-func TestExternalLocationListFails(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.1/unity-catalog/external-locations?",
-			Status:   200,
-			Response: &sdk_uc.ListExternalLocationsResponse{},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		err := resourcesMap["databricks_external_location"].List(ic)
-		assert.NoError(t, err)
-	})
-}
-
-func TestImportExternalLocationGrants(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Status:       200,
-			Resource:     "/api/2.1/unity-catalog/permissions/external-locations/abc",
-			Response: sdk_uc.GetPermissionsResponse{
-				PrivilegeAssignments: []sdk_uc.PrivilegeAssignment{
-					{
-						Principal:  "principal",
-						Privileges: []sdk_uc.Privilege{"ALL PRIVILEGES"},
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		d := tf_uc.ResourceExternalLocation().ToResource().TestResourceData()
-		d.SetId("abc")
-		err := resourcesMap["databricks_external_location"].Import(ic, &resource{
-			ID:   "abc",
-			Data: d,
-		})
-		assert.NoError(t, err)
-	})
-}
-
-// TODO: restore it when support for Account-level tests is added
-// func TestListMetastores(t *testing.T) {
-// 	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-// 		{
-// 			ReuseRequest: true,
-// 			Method:       "GET",
-// 			Resource:     "/api/2.1/unity-catalog/metastores",
-// 			Response: catalog.ListMetastoresResponse{
-// 				Metastores: []catalog.MetastoreInfo{
-// 					{
-// 						Name:        "test",
-// 						MetastoreId: "1234",
-// 					},
-// 				},
-// 			},
-// 		},
-// 	}, func(ctx context.Context, client *common.DatabricksClient) {
-// 		ic := importContextForTestWithClient(ctx, client)
-// 		ic.enableServices("uc-metastores")
-// 		err := resourcesMap["databricks_metastore"].List(ic)
-// 		assert.NoError(t, err)
-// 		require.Equal(t, 1, len(ic.testEmits))
-// 		assert.True(t, ic.testEmits["databricks_metastore[<unknown>] (id: 1234)"])
-// 	})
-// }
-
-func TestListCatalogs(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Resource:     "/api/2.1/unity-catalog/catalogs?",
-			Response: sdk_uc.ListCatalogsResponse{
-				Catalogs: []sdk_uc.CatalogInfo{
-					{
-						Name:        "cat1",
-						CatalogType: "MANAGED_CATALOG",
-					},
-					{
-						Name:        "cat2",
-						CatalogType: "UNKNOWN",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("uc-catalogs")
-		ic.currentMetastore = currentMetastoreResponse
-		err := resourcesMap["databricks_catalog"].List(ic)
-		assert.NoError(t, err)
-		require.Equal(t, 1, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_catalog[cat1_test_MANAGED_CATALOG] (id: cat1)"])
-	})
-}
-
-func TestImportManagedCatalog(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.1/unity-catalog/schemas?catalog_name=ctest",
-			Response: sdk_uc.ListSchemasResponse{
-				Schemas: []sdk_uc.SchemaInfo{
-					{
-						CatalogType: "MANAGED_CATALOG",
-						Name:        "schema1",
-						FullName:    "ctest.schema1",
-					},
-					{
-						CatalogType: "MANAGED_CATALOG",
-						Name:        "information_schema",
-						FullName:    "ctest.schema1",
-					},
-					{
-						CatalogType: "UNKNOWN",
-						FullName:    "ctest.schema2",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("uc-catalogs,uc-grants,uc-schemas")
-		ic.enableListing("uc-schemas")
-		ic.currentMetastore = currentMetastoreResponse
-		d := tf_uc.ResourceCatalog().ToResource().TestResourceData()
-		d.SetId("ctest")
-		d.Set("name", "ctest")
-		err := resourcesMap["databricks_catalog"].Import(ic, &resource{
-			ID:   "ctest",
-			Data: d,
-		})
-		assert.NoError(t, err)
-		require.Equal(t, 2, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: catalog/ctest)"])
-		assert.True(t, ic.testEmits["databricks_schema[<unknown>] (id: ctest.schema1)"])
-	})
-}
-
-func TestImportForeignCatalog(t *testing.T) {
-	ic := importContextForTest()
-	ic.enableServices("uc-catalogs,uc-grants,uc-connections")
-	ic.currentMetastore = currentMetastoreResponse
-	d := tf_uc.ResourceCatalog().ToResource().TestResourceData()
-	d.SetId("fctest")
-	d.Set("metastore_id", "1234")
-	d.Set("connection_name", "conn")
-	d.Set("name", "fctest")
-	err := resourcesMap["databricks_catalog"].Import(ic, &resource{
-		ID:   "fctest",
-		Data: d,
-	})
-	assert.NoError(t, err)
-	require.Equal(t, 2, len(ic.testEmits))
-	assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: catalog/fctest)"])
-	assert.True(t, ic.testEmits["databricks_connection[<unknown>] (id: 1234|conn)"])
-}
-
-func TestImportIsolatedManagedCatalog(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.1/unity-catalog/schemas?catalog_name=ctest",
-			Response: sdk_uc.ListSchemasResponse{},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.1/unity-catalog/bindings/catalog/ctest?",
-			Response: sdk_uc.UpdateWorkspaceBindingsResponse{
-				Bindings: []sdk_uc.WorkspaceBinding{
-					{
-						BindingType: "BINDING_TYPE_READ",
-						WorkspaceId: 1234,
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("uc-catalogs,uc-grants,uc-schemas")
-		ic.enableListing("uc-schemas,uc-volumes,uc-models,uc-tables")
-		ic.currentMetastore = currentMetastoreResponse
-		d := tf_uc.ResourceCatalog().ToResource().TestResourceData()
-		d.SetId("ctest")
-		d.Set("name", "ctest")
-		d.Set("isolation_mode", "ISOLATED")
-		err := resourcesMap["databricks_catalog"].Import(ic, &resource{
-			ID:   "ctest",
-			Data: d,
-		})
-		assert.NoError(t, err)
-		require.Equal(t, 2, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: catalog/ctest)"])
-		assert.True(t, ic.testEmits["databricks_workspace_binding[catalog_ctest_ws_1234] (id: 1234|catalog|ctest)"])
-	})
-}
-
-func TestImportSchema(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.1/unity-catalog/models?catalog_name=ctest&schema_name=stest",
-			Response: sdk_uc.ListRegisteredModelsResponse{
-				RegisteredModels: []sdk_uc.RegisteredModelInfo{
-					{
-						Name:     "model1",
-						FullName: "ctest.stest.model1",
-					},
-				},
-			},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.1/unity-catalog/volumes?catalog_name=ctest&schema_name=stest",
-			Response: sdk_uc.ListVolumesResponseContent{
-				Volumes: []sdk_uc.VolumeInfo{
-					{
-						Name:     "volume1",
-						FullName: "ctest.stest.volume1",
-					},
-				},
-			},
-		},
-		{
-			Method:   "GET",
-			Resource: "/api/2.1/unity-catalog/tables?catalog_name=ctest&schema_name=stest",
-			Response: sdk_uc.ListTablesResponse{
-				Tables: []sdk_uc.TableInfo{
-					{
-						Name:      "table1",
-						TableType: "MANAGED",
-						FullName:  "ctest.stest.table1",
-					},
-					{
-						Name:      "table2",
-						TableType: "UNKNOWN",
-						FullName:  "ctest.stest.table2",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("uc-catalogs,uc-grants,uc-schemas,uc-volumes,uc-models,uc-tables")
-		ic.enableListing("uc-schemas,uc-volumes,uc-models,uc-tables")
-		ic.currentMetastore = currentMetastoreResponse
-		d := tf_uc.ResourceSchema().ToResource().TestResourceData()
-		d.SetId("ctest.stest")
-		d.Set("catalog_name", "ctest")
-		d.Set("name", "stest")
-		err := resourcesMap["databricks_schema"].Import(ic, &resource{
-			ID:   "ctest.stest",
-			Data: d,
-		})
-		assert.NoError(t, err)
-		require.Equal(t, 5, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: schema/ctest.stest)"])
-		assert.True(t, ic.testEmits["databricks_catalog[<unknown>] (id: ctest)"])
-		assert.True(t, ic.testEmits["databricks_registered_model[<unknown>] (id: ctest.stest.model1)"])
-		assert.True(t, ic.testEmits["databricks_volume[<unknown>] (id: ctest.stest.volume1)"])
-		assert.True(t, ic.testEmits["databricks_sql_table[<unknown>] (id: ctest.stest.table1)"])
-	})
-}
-
-func TestImportShare(t *testing.T) {
-	ic := importContextForTest()
-	ic.enableServices("uc-grants,uc-volumes,uc-models,uc-tables")
-	d := tf_sharing.ResourceShare().ToResource().TestResourceData()
-	scm := tf_sharing.ResourceShare().Schema
-	share := tf_sharing.ShareInfo{
-		ShareInfo: sdk_sharing.ShareInfo{
-			Name: "stest",
-			Objects: []sdk_sharing.SharedDataObject{
-				{
-					DataObjectType: "TABLE",
-					Name:           "ctest.stest.table1",
-				},
-				{
-					DataObjectType: "MODEL",
-					Name:           "ctest.stest.model1",
-				},
-				{
-					DataObjectType: "VOLUME",
-					Name:           "ctest.stest.vol1",
-				},
-				{
-					DataObjectType: "NOTEBOOK",
-					Name:           "Test",
-				},
-			},
-		},
-	}
-	d.MarkNewResource()
-	err := common.StructToData(share, scm, d)
-	require.NoError(t, err)
-	err = resourcesMap["databricks_share"].Import(ic, &resource{
-		ID:   "stest",
-		Data: d,
-	})
-	assert.NoError(t, err)
-	require.Equal(t, 4, len(ic.testEmits))
-	assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: share/stest)"])
-	assert.True(t, ic.testEmits["databricks_registered_model[<unknown>] (id: ctest.stest.model1)"])
-	assert.True(t, ic.testEmits["databricks_volume[<unknown>] (id: ctest.stest.vol1)"])
-	assert.True(t, ic.testEmits["databricks_sql_table[<unknown>] (id: ctest.stest.table1)"])
-}
-
-func TestConnections(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Resource:     "/api/2.1/unity-catalog/connections?",
-			Response: sdk_uc.ListConnectionsResponse{
-				Connections: []sdk_uc.ConnectionInfo{
-					{
-						Name:        "test",
-						MetastoreId: "12345",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("uc-connections,uc-grants")
-		// Test Listing
-		err := resourcesMap["databricks_connection"].List(ic)
-		assert.NoError(t, err)
-		require.Equal(t, 1, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_connection[<unknown>] (id: 12345|test)"])
-		// Test Importing
-		d := tf_uc.ResourceConnection().ToResource().TestResourceData()
-		d.SetId("ctest")
-		d.Set("name", "ctest")
-		err = resourcesMap["databricks_connection"].Import(ic, &resource{
-			ID:   "ctest",
-			Data: d,
-		})
-		assert.NoError(t, err)
-		require.Equal(t, 2, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: foreign_connection/ctest)"])
-	})
-}
-
-func TestListExternalLocations(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Resource:     "/api/2.1/unity-catalog/external-locations?",
-			Response: sdk_uc.ListExternalLocationsResponse{
-				ExternalLocations: []sdk_uc.ExternalLocationInfo{
-					{
-						Name: "test",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("uc-external-locations,uc-storage-credentials,uc-grants")
-		ic.currentMetastore = currentMetastoreResponse
-		// Test listing
-		err := resourcesMap["databricks_external_location"].List(ic)
-		assert.NoError(t, err)
-		require.Equal(t, 1, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_external_location[<unknown>] (id: test)"])
-		// Test import
-		d := tf_uc.ResourceExternalLocation().ToResource().TestResourceData()
-		d.SetId("ext_loc")
-		d.Set("credential_name", "stest")
-		err = resourcesMap["databricks_external_location"].Import(ic, &resource{
-			ID:   "ext_loc",
-			Data: d,
-		})
-		assert.NoError(t, err)
-		require.Equal(t, 3, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: external_location/ext_loc)"])
-		assert.True(t, ic.testEmits["databricks_storage_credential[<unknown>] (id: stest)"])
-	})
-}
-
-func TestServiceCredentials(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Resource:     "/api/2.1/unity-catalog/credentials?",
-			Response: sdk_uc.ListCredentialsResponse{
-				Credentials: []sdk_uc.CredentialInfo{
-					{
-						Name:    "test-storage",
-						Purpose: sdk_uc.CredentialPurposeStorage,
-					},
-					{
-						Name:    "test-service",
-						Purpose: sdk_uc.CredentialPurposeService,
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("uc-credentials,uc-grants")
-		ic.currentMetastore = currentMetastoreResponse
-		// Test listing
-		err := resourcesMap["databricks_credential"].List(ic)
-		assert.NoError(t, err)
-		require.Equal(t, 1, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_credential[<unknown>] (id: test-service)"])
-		// Test import
-		err = resourcesMap["databricks_credential"].Import(ic, &resource{
-			ID: "1234",
-		})
-		assert.NoError(t, err)
-		require.Equal(t, 2, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: credential/1234)"])
-	})
-}
-
-func TestStorageCredentials(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Resource:     "/api/2.1/unity-catalog/storage-credentials?",
-			Response: sdk_uc.ListStorageCredentialsResponse{
-				StorageCredentials: []sdk_uc.StorageCredentialInfo{
-					{
-						Name: "test",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("uc-storage-credentials,uc-grants")
-		ic.currentMetastore = currentMetastoreResponse
-		// Test listing
-		err := resourcesMap["databricks_storage_credential"].List(ic)
-		assert.NoError(t, err)
-		require.Equal(t, 1, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_storage_credential[<unknown>] (id: test)"])
-		// Test import
-		err = resourcesMap["databricks_storage_credential"].Import(ic, &resource{
-			ID: "1234",
-		})
-		assert.NoError(t, err)
-		require.Equal(t, 2, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: storage_credential/1234)"])
-	})
-}
-
-func TestListRecipients(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Resource:     "/api/2.1/unity-catalog/recipients?",
-			Response: sdk_sharing.ListRecipientsResponse{
-				Recipients: []sdk_sharing.RecipientInfo{
-					{
-						Name: "test",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("uc-shares")
-		ic.currentMetastore = currentMetastoreResponse
-		err := resourcesMap["databricks_recipient"].List(ic)
-		assert.NoError(t, err)
-		require.Equal(t, 1, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_recipient[<unknown>] (id: test)"])
-	})
-}
-
-func TestVolumes(t *testing.T) {
-	ic := importContextForTest()
-	ic.enableServices("uc-volumes,uc-catalogs,uc-schemas,uc-grants")
-	// Test importing
-	d := tf_uc.ResourceVolume().ToResource().TestResourceData()
-	d.SetId("vtest")
-	d.Set("catalog_name", "ctest")
-	d.Set("schema_name", "stest")
-	err := resourcesMap["databricks_volume"].Import(ic, &resource{
-		ID:   "vtest",
-		Data: d,
-	})
-	assert.NoError(t, err)
-	require.Equal(t, 2, len(ic.testEmits))
-	assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: volume/vtest)"])
-	assert.True(t, ic.testEmits["databricks_schema[<unknown>] (id: ctest.stest)"])
-
-	//
-	r := &resource{
-		Attribute: "databricks_volume",
-		Value:     "dbc",
-		Data:      d,
-	}
-	shouldOmitFunc := resourcesMap["databricks_volume"].ShouldOmitField
-	require.NotNil(t, shouldOmitFunc)
-	scm := tf_uc.ResourceVolume().Schema
-	assert.False(t, shouldOmitFunc(nil, "volume_type", scm["volume_type"], d, r))
-	assert.False(t, shouldOmitFunc(nil, "name", scm["name"], d, r))
-	d.Set("volume_type", "MANAGED")
-	d.Set("storage_location", "s3://abc/")
-	assert.False(t, shouldOmitFunc(nil, "volume_type", scm["volume_type"], d, r))
-	assert.True(t, shouldOmitFunc(nil, "storage_location", scm["storage_location"], d, r))
-}
-
-func TestSqlTables(t *testing.T) {
-	ic := importContextForTest()
-	ic.enableServices("uc-tables,uc-catalogs,uc-schemas,uc-grants")
-	// Test importing
-	d := tf_uc.ResourceSqlTable().ToResource().TestResourceData()
-	d.SetId("ttest")
-	d.Set("catalog_name", "ctest")
-	d.Set("schema_name", "stest")
-	err := resourcesMap["databricks_sql_table"].Import(ic, &resource{
-		ID:   "ttest",
-		Data: d,
-	})
-	assert.NoError(t, err)
-	require.Equal(t, 2, len(ic.testEmits))
-	assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: table/ttest)"])
-	assert.True(t, ic.testEmits["databricks_schema[<unknown>] (id: ctest.stest)"])
-
-	//
-	r := &resource{
-		Attribute: "databricks_sql_table",
-		Value:     "ttest",
-		Data:      d,
-	}
-	shouldOmitFunc := resourcesMap["databricks_sql_table"].ShouldOmitField
-	require.NotNil(t, shouldOmitFunc)
-	scm := tf_uc.ResourceSqlTable().Schema
-	assert.False(t, shouldOmitFunc(nil, "table_type", scm["table_type"], d, r))
-	assert.False(t, shouldOmitFunc(nil, "name", scm["name"], d, r))
-	d.Set("table_type", "MANAGED")
-	d.Set("storage_location", "s3://abc/")
-	assert.False(t, shouldOmitFunc(nil, "table_type", scm["table_type"], d, r))
-	assert.True(t, shouldOmitFunc(nil, "storage_location", scm["storage_location"], d, r))
-}
-
-func TestRegisteredModels(t *testing.T) {
-	ic := importContextForTest()
-	ic.enableServices("uc-models,uc-catalogs,uc-schemas,uc-grants")
-	// Test importing
-	d := tf_uc.ResourceRegisteredModel().ToResource().TestResourceData()
-	d.SetId("mtest")
-	d.Set("catalog_name", "ctest")
-	d.Set("schema_name", "stest")
-	err := resourcesMap["databricks_registered_model"].Import(ic, &resource{
-		ID:   "mtest",
-		Data: d,
-	})
-	assert.NoError(t, err)
-	require.Equal(t, 2, len(ic.testEmits))
-	assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: model/mtest)"])
-	assert.True(t, ic.testEmits["databricks_schema[<unknown>] (id: ctest.stest)"])
-
-	//
-	r := &resource{
-		Attribute: "databricks_registered_model",
-		Value:     "mtest",
-		Data:      d,
-	}
-	shouldOmitFunc := resourcesMap["databricks_registered_model"].ShouldOmitField
-	require.NotNil(t, shouldOmitFunc)
-	scm := tf_uc.ResourceRegisteredModel().Schema
-	assert.True(t, shouldOmitFunc(nil, "storage_location", scm["storage_location"], d, r))
-	d.Set("storage_location", "s3://abc/")
-	assert.False(t, shouldOmitFunc(nil, "storage_location", scm["storage_location"], d, r))
-	assert.False(t, shouldOmitFunc(nil, "name", scm["name"], d, r))
-
-	ic.currentMetastore = currentMetastoreResponse
-	d.Set("storage_location", "s3://abc/"+currentMetastoreResponse.MetastoreId+"/models/123456")
-	assert.True(t, shouldOmitFunc(ic, "storage_location", scm["storage_location"], d, r))
-}
-
-func TestListShares(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			ReuseRequest: true,
-			Method:       "GET",
-			Resource:     "/api/2.1/unity-catalog/shares?",
-			Response: sdk_sharing.ListSharesResponse{
-				Shares: []sdk_sharing.ShareInfo{
-					{
-						Name: "test",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		ic.enableServices("uc-shares")
-		ic.currentMetastore = currentMetastoreResponse
-		err := resourcesMap["databricks_share"].List(ic)
-		assert.NoError(t, err)
-		require.Equal(t, 1, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_share[<unknown>] (id: test)"])
-	})
-}
-
-func TestAuxUcFunctions(t *testing.T) {
-	// Metastore Assignment
-	d := tf_uc.ResourceMetastoreAssignment().ToResource().TestResourceData()
-	d.Set("workspace_id", 123)
-	assert.Equal(t, "ws_123", resourcesMap["databricks_metastore_assignment"].Name(nil, d))
-	r := &resource{
-		Attribute: "databricks_metastore_assignment",
-		Value:     "ws_123",
-		Data:      d,
-	}
-
-	shouldOmitFunc := resourcesMap["databricks_metastore_assignment"].ShouldOmitField
-	require.NotNil(t, shouldOmitFunc)
-	d.Set("default_catalog_name", "")
-
-	scm := tf_uc.ResourceMetastoreAssignment().Schema
-	assert.True(t, shouldOmitFunc(nil, "default_catalog_name", scm["default_catalog_name"], d, r))
-	assert.False(t, shouldOmitFunc(nil, "metastore_id", scm["metastore_id"], d, r))
-
-	// Metastore
-	d = tf_uc.ResourceMetastore().ToResource().TestResourceData()
-	d.SetId("1234")
-	r = &resource{
-		Attribute: "databricks_metastore",
-		Value:     "1234",
-		Data:      d,
-	}
-	assert.Equal(t, "1234", resourcesMap["databricks_metastore"].Name(nil, d))
-	d.Set("name", "test")
-	assert.Equal(t, "test", resourcesMap["databricks_metastore"].Name(nil, d))
-
-	shouldOmitFunc = resourcesMap["databricks_metastore"].ShouldOmitField
-	require.NotNil(t, shouldOmitFunc)
-	scm = tf_uc.ResourceMetastore().Schema
-	assert.True(t, shouldOmitFunc(nil, "default_data_access_config_id", scm["default_data_access_config_id"], d, r))
-	assert.True(t, shouldOmitFunc(nil, "owner", scm["owner"], d, r))
-	d.Set("owner", "test")
-	assert.False(t, shouldOmitFunc(nil, "owner", scm["owner"], d, r))
-	assert.False(t, shouldOmitFunc(nil, "name", scm["name"], d, r))
-
-	// Connections
-	d = tf_uc.ResourceConnection().ToResource().TestResourceData()
-	d.SetId("1234")
-	assert.Equal(t, "1234", resourcesMap["databricks_connection"].Name(nil, d))
-	d.Set("name", "test")
-	d.Set("connection_type", "db")
-	assert.Equal(t, "db_test", resourcesMap["databricks_connection"].Name(nil, d))
-
-	// Catalogs
-	d = tf_uc.ResourceCatalog().ToResource().TestResourceData()
-	d.SetId("test")
-	r = &resource{
-		Attribute: "databricks_catalog",
-		Value:     "test",
-		Data:      d,
-	}
-	shouldOmitFunc = resourcesMap["databricks_catalog"].ShouldOmitField
-	require.NotNil(t, shouldOmitFunc)
-	scm = tf_uc.ResourceCatalog().Schema
-	d.Set("isolation_mode", "OPEN")
-	assert.True(t, shouldOmitFunc(nil, "isolation_mode", scm["isolation_mode"], d, r))
-	d.Set("isolation_mode", "ISOLATED")
-	assert.False(t, shouldOmitFunc(nil, "isolation_mode", scm["isolation_mode"], d, r))
-	assert.False(t, shouldOmitFunc(nil, "name", scm["name"], d, r))
 }
 
 func TestImportUcVolumeFile(t *testing.T) {
@@ -2431,343 +432,4 @@ func sortStringsCopy(s []string) []string {
 	copy(c, s)
 	sort.Strings(c)
 	return c
-}
-
-func TestImportGrants(t *testing.T) {
-	ic := importContextForTest()
-
-	s := ic.Resources["databricks_grants"].Schema
-	d := tf_uc.ResourceGrants().ToResource().TestResourceData()
-	id := "metastore/1234"
-	d.SetId(id)
-	d.MarkNewResource()
-	r := &resource{ID: id, Data: d}
-	err := resourcesMap["databricks_grants"].Import(ic, r)
-	assert.NoError(t, err)
-
-	var pList tf_uc.PermissionsList
-	common.DataToStructPointer(r.Data, s, &pList)
-	assert.Empty(t, pList.Assignments)
-
-	// Test with a filled user name and no owner
-	ic.meUserName = "user@domain.com"
-	d.Set("catalog", "1234")
-	err = resourcesMap["databricks_grants"].Import(ic, r)
-	assert.NoError(t, err)
-	common.DataToStructPointer(r.Data, s, &pList)
-	require.Equal(t, 0, len(pList.Assignments))
-
-	// Test with a filled user name and permissions
-	r.AddExtraData("owner", "otheruser@domain.com")
-	err = resourcesMap["databricks_grants"].Import(ic, r)
-	assert.NoError(t, err)
-	common.DataToStructPointer(r.Data, s, &pList)
-	require.Equal(t, 1, len(pList.Assignments))
-	assert.Equal(t, ic.meUserName, pList.Assignments[0].Principal)
-	assert.Equal(t, sortStringsCopy(grantsPrivilegesToAdd["catalog"]), sortStringsCopy(pList.Assignments[0].Privileges))
-
-	// Test with a filled user name and permissions
-	d.Set("metastore", "")
-	d.Set("catalog", "test")
-	pList.Assignments = []tf_uc.PrivilegeAssignment{
-		{Principal: ic.meUserName, Privileges: []string{"USE_CATALOG", "USE_SCHEMA"}},
-	}
-	common.StructToData(pList, s, d)
-	err = resourcesMap["databricks_grants"].Import(ic, r)
-	assert.NoError(t, err)
-	common.DataToStructPointer(r.Data, s, &pList)
-	require.Equal(t, 1, len(pList.Assignments))
-	assert.Equal(t, ic.meUserName, pList.Assignments[0].Principal)
-	assert.Equal(t, sortStringsCopy(append([]string{"USE_CATALOG", "USE_SCHEMA"}, grantsPrivilegesToAdd["catalog"]...)),
-		sortStringsCopy(pList.Assignments[0].Privileges))
-
-	// Test with a filled user name and unsupported objects
-	d.Set("catalog", "")
-	d.Set("model", "test")
-	err = resourcesMap["databricks_grants"].Import(ic, r)
-	assert.NoError(t, err)
-}
-
-func TestImportUcOnlineTable(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		tmpDir := fmt.Sprintf("/tmp/tf-%s", qa.RandomName())
-		defer os.RemoveAll(tmpDir)
-		os.Mkdir(tmpDir, 0700)
-		ic.Directory = tmpDir
-		ic.enableServices("uc-tables,uc-grants")
-		ic.currentMetastore = currentMetastoreResponse
-
-		otTableName := "main.tmp.tbl_ot"
-		d := tf_uc.ResourceOnlineTable().ToResource().TestResourceData()
-		ot := sdk_uc.OnlineTable{
-			Name: otTableName,
-			Spec: &sdk_uc.OnlineTableSpec{
-				SourceTableFullName: "main.tmp.tbl",
-				PrimaryKeyColumns:   []string{"id"},
-			},
-		}
-		d.SetId(otTableName)
-		d.MarkNewResource()
-		scm := tf_uc.ResourceOnlineTable().Schema
-		err := common.StructToData(ot, scm, d)
-		require.NoError(t, err)
-
-		err = resourcesMap["databricks_online_table"].Import(ic, &resource{
-			ID:   otTableName,
-			Data: d,
-		})
-		assert.NoError(t, err)
-		require.Equal(t, 2, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_sql_table[<unknown>] (id: main.tmp.tbl)"])
-		assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: table/main.tmp.tbl_ot)"])
-	})
-}
-
-func TestImportVectorSearchEndpointList(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/vector-search/endpoints?",
-			Response: sdk_vs.ListEndpointResponse{
-				Endpoints: []sdk_vs.EndpointInfo{
-					{
-						Name:                 "test",
-						LastUpdatedTimestamp: 1234567890,
-					},
-					{
-						Name:                 "test2",
-						LastUpdatedTimestamp: 2234567890,
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		tmpDir := fmt.Sprintf("/tmp/tf-%s", qa.RandomName())
-		defer os.RemoveAll(tmpDir)
-		os.Mkdir(tmpDir, 0700)
-		ic.Directory = tmpDir
-		ic.enableServices("vector-search")
-		ic.currentMetastore = currentMetastoreResponse
-
-		err := resourcesMap["databricks_vector_search_endpoint"].List(ic)
-		assert.NoError(t, err)
-		require.Equal(t, 2, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_vector_search_endpoint[<unknown>] (id: test)"])
-		assert.True(t, ic.testEmits["databricks_vector_search_endpoint[<unknown>] (id: test2)"])
-	})
-}
-
-func TestImportVectorSearchEndpoint(t *testing.T) {
-	vseName := "test"
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/vector-search/indexes?endpoint_name=test",
-			Response: sdk_vs.ListVectorIndexesResponse{
-				VectorIndexes: []sdk_vs.MiniVectorIndex{
-					{
-						Name: "idx1",
-					},
-					{
-						Name: "idx2",
-					},
-				},
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		tmpDir := fmt.Sprintf("/tmp/tf-%s", qa.RandomName())
-		defer os.RemoveAll(tmpDir)
-		os.Mkdir(tmpDir, 0700)
-		ic.Directory = tmpDir
-		ic.enableServices("vector-search")
-		ic.currentMetastore = currentMetastoreResponse
-
-		d := tf_vs.ResourceVectorSearchEndpoint().ToResource().TestResourceData()
-		vse := sdk_vs.EndpointInfo{
-			Name: vseName,
-		}
-		d.SetId(vseName)
-		d.MarkNewResource()
-		scm := tf_vs.ResourceVectorSearchEndpoint().Schema
-		err := common.StructToData(vse, scm, d)
-		require.NoError(t, err)
-
-		err = resourcesMap["databricks_vector_search_endpoint"].Import(ic, &resource{
-			ID:   vseName,
-			Data: d,
-		})
-		assert.NoError(t, err)
-		require.Equal(t, 2, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_vector_search_index[<unknown>] (id: idx1)"])
-		assert.True(t, ic.testEmits["databricks_vector_search_index[<unknown>] (id: idx2)"])
-	})
-}
-
-func TestImportVectorSearchIndex(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTestWithClient(ctx, client)
-		tmpDir := fmt.Sprintf("/tmp/tf-%s", qa.RandomName())
-		defer os.RemoveAll(tmpDir)
-		os.Mkdir(tmpDir, 0700)
-		ic.Directory = tmpDir
-		ic.enableServices("vector-search,uc-tables,uc-grants,model-serving")
-		ic.currentMetastore = currentMetastoreResponse
-
-		vsiName := "main.tmp.vsi"
-		d := tf_vs.ResourceVectorSearchIndex().ToResource().TestResourceData()
-		ot := sdk_vs.VectorIndex{
-			Name:         vsiName,
-			PrimaryKey:   "id",
-			EndpointName: "vs",
-			DeltaSyncIndexSpec: &sdk_vs.DeltaSyncVectorIndexSpecResponse{
-				SourceTable: "main.tmp.tbl",
-				EmbeddingSourceColumns: []sdk_vs.EmbeddingSourceColumn{
-					{
-						Name:                       "col1",
-						EmbeddingModelEndpointName: "test",
-					},
-				},
-			},
-			DirectAccessIndexSpec: &sdk_vs.DirectAccessVectorIndexSpec{
-				EmbeddingSourceColumns: []sdk_vs.EmbeddingSourceColumn{
-					{
-						Name:                       "col1",
-						EmbeddingModelEndpointName: "test",
-					},
-				},
-			},
-		}
-		d.SetId(vsiName)
-		d.MarkNewResource()
-		scm := tf_vs.ResourceVectorSearchIndex().Schema
-		err := common.StructToData(ot, scm, d)
-		require.NoError(t, err)
-
-		err = resourcesMap["databricks_vector_search_index"].Import(ic, &resource{
-			ID:   vsiName,
-			Data: d,
-		})
-		assert.NoError(t, err)
-		require.Equal(t, 4, len(ic.testEmits))
-		assert.True(t, ic.testEmits["databricks_grants[<unknown>] (id: table/main.tmp.vsi)"])
-		assert.True(t, ic.testEmits["databricks_vector_search_endpoint[<unknown>] (id: vs)"])
-		assert.True(t, ic.testEmits["databricks_sql_table[<unknown>] (id: main.tmp.tbl)"])
-		assert.True(t, ic.testEmits["databricks_model_serving[<unknown>] (id: test)"])
-	})
-}
-
-func TestImportVectorSearchIndexWithDirectAccess(t *testing.T) {
-	d := tf_vs.ResourceVectorSearchIndex().ToResource().TestResourceData()
-	d.SetId("test-index")
-	d.Set("endpoint_name", "test-endpoint")
-
-	// Set up direct access index spec
-	directAccess := map[string]any{
-		"embedding_source_columns": []map[string]any{
-			{
-				"name":                          "embedding_column",
-				"embedding_model_endpoint_name": "embedding-model",
-			},
-		},
-	}
-	d.Set("direct_access_index_spec", []map[string]any{directAccess})
-
-	ic := importContextForTest()
-	ic.enableServices("vector-search,uc-tables,uc-grants,model-serving")
-
-	r := &resource{
-		Resource: "databricks_vector_search_index",
-		ID:       "test-index",
-		Data:     d,
-	}
-
-	err := resourcesMap["databricks_vector_search_index"].Import(ic, r)
-	assert.NoError(t, err)
-
-	// Check that the right resources were emitted
-	assert.True(t, ic.testEmits["databricks_vector_search_endpoint[<unknown>] (id: test-endpoint)"])
-	assert.True(t, ic.testEmits["databricks_model_serving[<unknown>] (id: embedding-model)"])
-}
-
-func TestNotificationDestinationName(t *testing.T) {
-	ic := importContextForTest()
-	d := tf_settings.ResourceNotificationDestination().ToResource().TestResourceData()
-
-	// Test with display name
-	d.SetId("abcdef1234567890")
-	d.Set("display_name", "Test Notification")
-	name := resourcesMap["databricks_notification_destination"].Name(ic, d)
-	assert.Equal(t, "Test_Notification__abcdef12", name)
-
-	// Test without display name
-	d.Set("display_name", "")
-	name = resourcesMap["databricks_notification_destination"].Name(ic, d)
-	assert.Equal(t, "_abcdef12", name)
-}
-
-func TestNotificationDestinationImport(t *testing.T) {
-	d := tf_settings.ResourceNotificationDestination().ToResource().TestResourceData()
-	d.SetId("test-notification")
-	d.Set("display_name", "Test Notification")
-	d.Set("destination_type", "EMAIL")
-
-	// Set email config
-	emailConfig := map[string]any{
-		"addresses": []string{"user@example.com"},
-	}
-
-	config := map[string]any{
-		"email": []map[string]any{emailConfig},
-	}
-	d.Set("config", []map[string]any{config})
-
-	ic := importContextForTest()
-	r := &resource{
-		Resource: "databricks_notification_destination",
-		ID:       "test-notification",
-		Data:     d,
-	}
-
-	err := resourcesMap["databricks_notification_destination"].Import(ic, r)
-	assert.NoError(t, err)
-}
-
-func TestNotificationDestinationShouldOmitField(t *testing.T) {
-	d := tf_settings.ResourceNotificationDestination().ToResource().TestResourceData()
-	d.SetId("test-notification")
-
-	// Test webhook configuration
-	d.Set("destination_type", "WEBHOOK")
-
-	// Set webhook config with URL not set
-	webhookConfig := map[string]any{
-		"url":          "https://example.com/webhook",
-		"url_set":      false,
-		"username":     "webhook-user",
-		"username_set": true,
-	}
-
-	config := map[string]any{
-		"generic_webhook": []map[string]any{webhookConfig},
-	}
-	d.Set("config", []map[string]any{config})
-
-	ic := importContextForTest()
-	schema := tf_settings.ResourceNotificationDestination().Schema
-	r := &resource{
-		Attribute: "databricks_notification_destination",
-		Value:     "test-notification",
-		Data:      d,
-	}
-
-	// URL should be omitted because url_set is false
-	assert.True(t, resourcesMap["databricks_notification_destination"].ShouldOmitField(
-		ic, "config.0.generic_webhook.0.url", schema["config"], d, r))
-
-	// Username should not be omitted because username_set is true
-	assert.False(t, resourcesMap["databricks_notification_destination"].ShouldOmitField(
-		ic, "config.0.generic_webhook.0.username", schema["config"], d, r))
 }
